New code to handle read timeout when waiting for a result from the HSM
- Speed improvment with 200-300%, going from 30-40ms to <= 2ms per command

+++ src/main/java/org/unitedid/yhsm/YubiHSM.java
@@ -43,7 +43,7 @@ public class YubiHSM  {
      * @param timeout the read timeout (not implemented yet)
      * @throws YubiHSMErrorException if the YubiHSM reset command fail
      */
-    public YubiHSM(String device, int timeout) throws YubiHSMErrorException {
+    public YubiHSM(String device, float timeout) throws YubiHSMErrorException {
         deviceHandler = DeviceHandlerFactory.get(device, timeout);
         CommandHandler.reset(deviceHandler);
     }
